AI-Cold-Outreach-Engine/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-serpapi
python-dotenv
pandas
